+++ gpu4pyscf/df/df_jk.py
@@ -1,4 +1,4 @@
-# Copyright 2021-2024 The PySCF Developers. All Rights Reserved.
+# Copyright 2021-2026 The PySCF Developers. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -20,18 +20,19 @@ from concurrent.futures import ThreadPoolExecutor
 import cupy
 import numpy
 import cupy as cp
-from pyscf import lib, __config__
+from pyscf import lib
 from pyscf.scf import dhf
 from gpu4pyscf.lib import logger
 from gpu4pyscf.lib.cupy_helper import (
-    contract, transpose_sum, reduce_to_device, tag_array, CPArrayWithTag)
+    contract, transpose_sum, tag_array, CPArrayWithTag, cholesky)
 from gpu4pyscf.dft import rks, uks, numint, gks
 from gpu4pyscf.scf import hf, uhf, rohf
 from gpu4pyscf.scf.jk import _check_rsh_factors
-from gpu4pyscf.df import df, int3c2e
-from gpu4pyscf.__config__ import num_devices
+from gpu4pyscf.df import df
 from gpu4pyscf.scf import ghf
-from gpu4pyscf.lib.cupy_helper import asarray
+from gpu4pyscf.lib.cupy_helper import asarray, ndarray, get_avail_mem
+from gpu4pyscf.lib import multi_gpu
+num_devices = multi_gpu.num_devices
 
 def _density_fit(mf, auxbasis=None, with_df=None, only_dfj=False):
     '''For the given SCF object, update the J, K matrix constructor with
@@ -119,12 +120,12 @@ class _DFHF:
     def get_k(self, mol=None, dm=None, hermi=1, omega=None,
               lr_factor=None, sr_factor=None):
         omega, lr_factor, sr_factor = _check_rsh_factors(mol, omega, lr_factor, sr_factor)
-        vk = self.with_df.get_jk(dm, hermi, False, True, self.direct_scf_tol)[1]
+        vk = self.with_df.get_jk(dm, hermi, False, True)[1]
         vk *= sr_factor
         if omega == 0:
             return vk
 
-        vklr = self.with_df.get_jk(dm, hermi, False, True, self.direct_scf_tol, omega)[1]
+        vklr = self.with_df.get_jk(dm, hermi, False, True, omega=omega)[1]
         vklr *= lr_factor - sr_factor
         vk += vklr
         return vk
@@ -142,9 +143,7 @@ class _DFHF:
                     def jkbuild(mol_obj, dm_obj, hermi, omega=None):
                         nao = mol_obj.nao
                         dm_obj = dm_obj.reshape(-1, nao, nao)
-                        return self.with_df.get_jk(dm_obj, hermi,
-                                                   direct_scf_tol=self.direct_scf_tol,
-                                                   omega=omega)
+                        return self.with_df.get_jk(dm_obj, hermi, omega=omega)
                     # Pass with_k=False to get only DF J
                     vj, _ = ghf.get_jk(mol, dm, hermi, True, False,
                                        jkbuild=jkbuild, omega=omega)
@@ -163,16 +162,13 @@ class _DFHF:
                 def jkbuild(mol_obj, dm_obj, hermi, omega=None):
                     nao = mol_obj.nao
                     dm_obj = dm_obj.reshape(-1, nao, nao)
-                    return self.with_df.get_jk(dm_obj, hermi,
-                                               direct_scf_tol=self.direct_scf_tol,
-                                               omega=omega)
+                    return self.with_df.get_jk(dm_obj, hermi, omega=omega)
                 vj, vk = ghf.get_jk(mol, dm, hermi, with_j, with_k,
                                     jkbuild=jkbuild, omega=omega)
             else:
                 # Standard RHF/UHF: Use Master's direct call
-                vj, vk = self.with_df.get_jk(dm, hermi, with_j, with_k,
-                                             self.direct_scf_tol, omega)
-                                             
+                vj, vk = self.with_df.get_jk(dm, hermi, with_j, with_k, omega=omega)
+
         else:
             # Error handling from Master
             raise ValueError(f"with_df field not found in a df object (type = {type(self)}) during a get_jk() call.")
@@ -321,8 +317,6 @@ class _DFHF:
                     exc -= float(cupy.einsum('ij,ji->', dm, vk).real.get()) * .25
                 ecoul = float(cupy.einsum('ij,ji->', dm, vj).real.get()) * .5
             elif isinstance(self, ghf.GHF):
-                ground_state = isinstance(dm, cupy.ndarray) and dm.ndim == 2
-                
                 if hermi == 2:  # because rho = 0
                     n, exc, vxc = 0, 0, 0
                 else:
@@ -375,16 +369,9 @@ class _DFHF:
                         vklr = self.get_k(mol, dm, hermi, omega=omega)
                         vklr *= (alpha - hyb)
                         vk += vklr
-                    
                     vxc += vj - vk
-
-                    if ground_state:
-                        exc -= cupy.einsum('ij,ji', dm, vk).real * .5
-
-                if ground_state:
-                    ecoul = cupy.einsum('ij,ji', dm, vj).real * .5
-                else:
-                    ecoul = None
+                    exc -= cupy.einsum('ij,ji->', dm, vk).real * .5
+                    ecoul = cupy.einsum('ij,ji->', dm, vj).real * .5
             else:
                 raise NotImplementedError("DF only supports R/U/RO KS.")
             t0 = log.timer('veff', *t0)
@@ -401,7 +388,7 @@ class _DFHF:
             vhf = vj - vk * .5
             ecoul = float(cp.einsum('ij,ji->', dm, vj).real.get()) * .5
             return tag_array(vhf, ecoul=ecoul)
-        elif isinstance(self, ghf.GHF): # (New) GHF branch
+        elif isinstance(self, ghf.GHF): # GHF branch
             vj, vk = self.get_jk(mol, dm, hermi=hermi)
             vhf = vj - vk
             ecoul = float(cp.einsum('ij,ji->', dm, vj).real.get()) * .5
@@ -418,393 +405,157 @@ class _DFHF:
             obj.spin_samples = self.spin_samples
         return obj
 
-def _jk_task_with_mo(dfobj, dms, mo_coeff, mo_occ,
-                     with_j=True, with_k=True, hermi=0, device_id=0):
-    ''' Calculate J and K matrices on single GPU
-    '''
-    with cupy.cuda.Device(device_id):
-        assert isinstance(dfobj.verbose, int)
-        log = logger.new_logger(dfobj.mol, dfobj.verbose)
-        t0 = log.init_timer()
-        dms = cupy.asarray(dms, order='A')
-        mo_coeff = cupy.asarray(mo_coeff)
-        mo_occ = cupy.asarray(mo_occ)
-        nao = dms.shape[-1]
-        intopt = dfobj.intopt
-        rows = intopt.cderi_row
-        cols = intopt.cderi_col
-        nset = dms.shape[0]
-        dms_shape = dms.shape
-        vj = vk = None
-        if with_j:
-            dm_sparse = dms[:,rows,cols]
-            if hermi == 0:
-                dm_sparse += dms[:,cols,rows]
-            else:
-                dm_sparse *= 2
-            dm_sparse[:, intopt.cderi_diag] *= .5
-
-        if with_k:
-            vk = cupy.zeros_like(dms)
-
-        # SCF K matrix with occ
-        if mo_coeff is not None:
-            assert hermi == 1
-            nocc = 0
-            occ_coeff = [0]*nset
-            for i in range(nset):
-                occ_idx = mo_occ[i] > 0
-                occ_coeff[i] = mo_coeff[i][:,occ_idx] * mo_occ[i][occ_idx]**0.5
-                nocc += int(mo_occ[i].sum())
-            blksize = dfobj.get_blksize(extra=nao*nocc)
-            if with_j:
-                vj_packed = cupy.zeros_like(dm_sparse)
-            for cderi, cderi_sparse in dfobj.loop(blksize=blksize, unpack=with_k):
-                # leading dimension is 1
-                if with_j:
-                    rhoj = dm_sparse.dot(cderi_sparse)
-                    vj_packed += cupy.dot(rhoj, cderi_sparse.T)
-                cderi_sparse = rhoj = None
-                for i in range(nset):
-                    if with_k:
-                        rhok = contract('Lji,jk->Lki', cderi, occ_coeff[i])
-                        # In most cases, syrk does not outperform cupy.dot
-                        #cublas.syrk('T', rhok.reshape([-1,nao]), out=vk[i], alpha=1.0, beta=1.0, lower=True)
-                        rhok = rhok.reshape([-1,nao])
-                        vk[i] += cupy.dot(rhok.T, rhok)
-                    rhok = None
-
-            if with_j:
-                vj = cupy.zeros(dms_shape)
-                vj[:,rows,cols] = vj_packed
-                vj[:,cols,rows] = vj_packed
-        t0 = log.timer_debug1(f'vj and vk on Device {device_id}', *t0)
-    return vj, vk
-
-def _jk_task_with_mo1(dfobj, dms, mo1s, occ_coeffs,
-                      with_j=True, with_k=True, hermi=0, device_id=0):
-    ''' Calculate J and K matrices with mo response
-        For CP-HF or TDDFT
-    '''
-    vj = vk = None
-    with cupy.cuda.Device(device_id):
-        assert isinstance(dfobj.verbose, int)
-        log = logger.new_logger(dfobj.mol, dfobj.verbose)
-        t0 = log.init_timer()
-        dms = cupy.asarray(dms, order='A')
-        mo1s = [cupy.asarray(mo1, order='A') for mo1 in mo1s]
-        occ_coeffs = [cupy.asarray(occ_coeff, order='A') for occ_coeff in occ_coeffs]
-
-        nao = dms.shape[-1]
-        intopt = dfobj.intopt
-        rows = intopt.cderi_row
-        cols = intopt.cderi_col
-        dms_shape = dms.shape
-        if with_j:
-            dm_sparse = dms[:,rows,cols]
-            if hermi == 0:
-                dm_sparse += dms[:,cols,rows]
-            else:
-                dm_sparse *= 2
-            dm_sparse[:, intopt.cderi_diag] *= .5
-
-        if with_k:
-            vk = cupy.zeros_like(dms)
-
-        if with_j:
-            vj_sparse = cupy.zeros_like(dm_sparse)
-
-        nocc = max([mo1.shape[2] for mo1 in mo1s])
-        blksize = dfobj.get_blksize(extra=2*nao*nocc)
-        for cderi, cderi_sparse in dfobj.loop(blksize=blksize, unpack=with_k):
-            if with_j:
-                rhoj = dm_sparse.dot(cderi_sparse)
-                vj_sparse += cupy.dot(rhoj, cderi_sparse.T)
-                rhoj = None
-            cderi_sparse = None
-            if with_k:
-                iset = 0
-                for occ_coeff, mo1 in zip(occ_coeffs, mo1s):
-                    rhok = contract('Lij,jk->Lki', cderi, occ_coeff).reshape([-1,nao])
-                    for i in range(mo1.shape[0]):
-                        rhok1 = contract('Lij,jk->Lki', cderi, mo1[i]).reshape([-1,nao])
-                        #contract('Lki,Lkj->ij', rhok1, rhok, alpha=1.0, beta=1.0, out=vk[iset])
-                        vk[iset] += cupy.dot(rhok1.T, rhok)
-                        iset += 1
-                mo1 = rhok1 = rhok = None
-            cderi = None
-        mo1s = None
-        if with_j:
-            vj = cupy.zeros(dms_shape)
-            vj[:,rows,cols] = vj_sparse
-            vj[:,cols,rows] = vj_sparse
-        if with_k and hermi:
-            transpose_sum(vk)
-        vj_sparse = None
-
-        t0 = log.timer_debug1(f'vj and vk on Device {device_id}', *t0)
-    return vj, vk
-
-def _jk_via_decomposed_dm(dfobj, dms, hermi=0, with_j=True, with_k=True, device_id=0):
-    with cupy.cuda.Device(device_id):
-        nao = dms.shape[-1]
-        intopt = dfobj.intopt
-        # dms = symmetrize(factor_l.dot(factor_r.T))
-        symmetrize = getattr(dms, 'symmetrize', 0)
-        dm_factor_l = dms.factor_l
-        dm_factor_r = dms.factor_r
-        dm_factor_l = cp.asarray(dm_factor_l, order='A')
-        dm_factor_l = intopt.sort_orbitals(dm_factor_l, axis=[dm_factor_l.ndim-2])
-        if dm_factor_r is None:
-            dm_factor_r = dm_factor_l
-        else:
-            dm_factor_r = cp.asarray(dm_factor_r, order='A')
-            dm_factor_r = intopt.sort_orbitals(dm_factor_r, axis=[dm_factor_r.ndim-2])
-        nocc = dm_factor_l.shape[-1]
-        dms = dms.reshape(-1,nao,nao)
-        n_dm = len(dms)
-        if dm_factor_l.ndim == dm_factor_r.ndim:
-            dm_factor_l = dm_factor_l.reshape(n_dm, nao, nocc)
-            dm_factor_r = dm_factor_r.reshape(n_dm, nao, nocc)
-
-        vj = vk = None
-        if with_j:
-            rows = cp.asarray(intopt.cderi_row)
-            cols = cp.asarray(intopt.cderi_col)
-            dms = cp.asarray(dms, order='A')
-            dms = intopt.sort_orbitals(dms, axis=[1,2])
-            dm_sparse = dms[:,rows,cols]
-            if hermi == 0:
-                dm_sparse += dms[:,cols,rows]
-            else:
-                dm_sparse *= 2
-            dm_sparse[:, intopt.cderi_diag] *= .5
-            vj_sparse = cupy.zeros_like(dm_sparse)
-
-        if with_k:
-            vk = cupy.zeros_like(dms)
-
-        blksize = dfobj.get_blksize(extra=2*nao*nocc)
-        buf1 = cp.empty((nao, nao))
-        buf2 = cp.empty((blksize, nocc, nao))
-        buf3 = cp.empty((blksize, nocc, nao))
-        for cderi, cderi_sparse in dfobj.loop(blksize=blksize, unpack=with_k):
-            if with_j:
-                rhoj = dm_sparse.dot(cderi_sparse)
-                vj_sparse += cupy.dot(rhoj, cderi_sparse.T)
-                rhoj = None
-            cderi_sparse = None
-            if with_k:
-                nL = len(cderi)
-                rhok = buf2[:nL]
-                rhok1 = buf3[:nL]
-                if dm_factor_l.ndim == dm_factor_r.ndim:
-                    for i in range(n_dm):
-                        contract('Lij,jk->Lki', cderi, dm_factor_l[i], out=rhok)
-                        contract('Lij,jk->Lki', cderi, dm_factor_r[i], out=rhok1)
-                        vk[i] += cupy.dot(rhok.reshape([-1,nao]).T,
-                                          rhok1.reshape([-1,nao]), out=buf1)
-                elif dm_factor_l.ndim < dm_factor_r.ndim:
-                    contract('Lij,jk->Lki', cderi, dm_factor_l, out=rhok)
-                    for i in range(n_dm):
-                        contract('Lij,jk->Lki', cderi, dm_factor_r[i], out=rhok1)
-                        vk[i] += cupy.dot(rhok.reshape([-1,nao]).T,
-                                          rhok1.reshape([-1,nao]), out=buf1)
-                else:
-                    contract('Lij,jk->Lki', cderi, dm_factor_r, out=rhok1)
-                    for i in range(n_dm):
-                        contract('Lij,jk->Lki', cderi, dm_factor_l[i], out=rhok)
-                        vk[i] += cupy.dot(rhok.reshape([-1,nao]).T,
-                                          rhok1.reshape([-1,nao]), out=buf1)
-                rhok1 = rhok = None
-            cderi = None
-
-        if with_j:
-            vj = cupy.zeros(dms.shape)
-            vj[:,rows,cols] = vj_sparse
-            vj[:,cols,rows] = vj_sparse
-        if with_k:
-            if symmetrize != 0:
-                vk = transpose_sum(vk, hermi=symmetrize)
-    return vj, vk
-
-def _jk_task_with_dm(dfobj, dms, with_j=True, with_k=True, hermi=0, device_id=0):
-    ''' Calculate J and K matrices with density matrix
-    '''
-    with cupy.cuda.Device(device_id):
-        assert isinstance(dfobj.verbose, int)
-        log = logger.new_logger(dfobj.mol, dfobj.verbose)
-        t0 = log.init_timer()
-        dms = cupy.asarray(dms, order='A')
-        intopt = dfobj.intopt
-        rows = intopt.cderi_row
-        cols = intopt.cderi_col
-        nao = dms.shape[-1]
-        dms_shape = dms.shape
-        vj = vk = None
-        if with_j:
-            dm_sparse = dms[:,rows,cols]
-            if hermi == 0:
-                dm_sparse += dms[:,cols,rows]
-            else:
-                dm_sparse *= 2
-            dm_sparse[:, intopt.cderi_diag] *= .5
-            vj_sparse = cupy.zeros_like(dm_sparse)
-
-        if with_k:
-            vk = cupy.zeros_like(dms)
-
-        nset = dms.shape[0]
-        blksize = dfobj.get_blksize()
-        for cderi, cderi_sparse in dfobj.loop(blksize=blksize, unpack=with_k):
-            if with_j:
-                rhoj = dm_sparse.dot(cderi_sparse)
-                vj_sparse += cupy.dot(rhoj, cderi_sparse.T)
-            if with_k:
-                for k in range(nset):
-                    rhok = contract('Lij,jk->Lki', cderi, dms[k]).reshape([-1,nao])
-                    #vk[k] += contract('Lki,Lkj->ij', rhok, cderi)
-                    vk[k] += cupy.dot(rhok.T, cderi.reshape([-1,nao]))
-        if with_j:
-            vj = cupy.zeros(dms_shape)
-            vj[:,rows,cols] = vj_sparse
-            vj[:,cols,rows] = vj_sparse
-
-        t0 = log.timer_debug1(f'vj and vk on Device {device_id}', *t0)
-    return vj, vk
-
-def get_jk(dfobj, dms_tag, hermi=0, with_j=True, with_k=True, direct_scf_tol=1e-14, omega=None):
+def get_jk(dfobj, dms, hermi=0, with_j=True, with_k=True, omega=None):
     '''
     get jk with density fitting
-    outputs and input are on the same device
-    TODO: separate into three cases: j only, k only, j and k
     '''
-
     log = logger.new_logger(dfobj.mol, dfobj.verbose)
-    out_shape = dms_tag.shape
-    out_cupy = isinstance(dms_tag, cupy.ndarray)
-    if not isinstance(dms_tag, cupy.ndarray):
-        dms_tag = cupy.asarray(dms_tag)
-
-    assert(with_j or with_k)
-    if dms_tag is None: logger.error("dm is not given")
-    nao = dms_tag.shape[-1]
     t1 = t0 = log.init_timer()
     if dfobj._cderi is None:
         log.debug('Build CDERI ...')
-        dfobj.build(direct_scf_tol=direct_scf_tol, omega=omega)
+        dfobj.build(omega=omega)
         t1 = log.timer_debug1('init jk', *t0)
 
-    assert nao == dfobj.nao
-    intopt = dfobj.intopt
+    out_cupy = isinstance(dms, cp.ndarray)
+    dm_factor_l, dm_factor_r = factorize_dm(dms, hermi)
+    symmetrize = getattr(dms, 'symmetrize', 0)
 
-    nao = dms_tag.shape[-1]
-    dms = dms_tag.reshape([-1,nao,nao])
-    intopt = dfobj.intopt
+    nao, nocc = dm_factor_l.shape[-2:]
+    dms_3d = cp.asarray(dms).reshape(-1,nao,nao)
+    n_dm = dms_3d.shape[0]
 
-    if getattr(dms_tag, 'mo_coeff', None) is not None:
-        dms = intopt.sort_orbitals(dms, axis=[1,2])
-        mo_occ = dms_tag.mo_occ
-        mo_coeff = dms_tag.mo_coeff
-        nmo = mo_occ.shape[-1]
-        mo_coeff = mo_coeff.reshape(-1,nao,nmo)
-        mo_occ   = mo_occ.reshape(-1,nmo)
-        mo_coeff = intopt.sort_orbitals(mo_coeff, axis=[1])
-        cupy.cuda.get_current_stream().synchronize()
+    def proc():
+        factor_l = cp.asarray(dm_factor_l)
+        factor_r = dm_factor_r
+        if factor_r is not None:
+            factor_r = cp.asarray(factor_r)
 
-        futures = []
-        with ThreadPoolExecutor(max_workers=num_devices) as executor:
-            for device_id in range(num_devices):
-                future = executor.submit(
-                    _jk_task_with_mo,
-                    dfobj, dms, mo_coeff, mo_occ,
-                    hermi=hermi, device_id=device_id,
-                    with_j=with_j, with_k=with_k)
-                futures.append(future)
+        vj = vk = None
+        if with_j:
+            pair_addresses, diags = dfobj._cderi_idx
+            rows, cols = divmod(cp.asarray(pair_addresses), nao)
+            dm_sparse = cp.asarray(dms_3d)[:,rows,cols]
+            if hermi == 0:
+                dm_sparse += cp.asarray(dms_3d)[:,cols,rows]
+            else:
+                dm_sparse *= 2
+            dm_sparse[:,cp.asarray(diags)] *= .5
+            vj_sparse = cp.zeros_like(dm_sparse)
 
-    elif hasattr(dms_tag, 'mo1'):
-        dms = intopt.sort_orbitals(dms, axis=[1,2])
-        occ_coeffs = dms_tag.occ_coeff
-        mo1s = dms_tag.mo1
-        if not isinstance(occ_coeffs, (tuple, list)):
-            # *2 for double occupancy in RHF/RKS
-            occ_coeffs = [occ_coeffs * 2.0]
-        if not isinstance(mo1s, (tuple, list)):
-            mo1s = [mo1s]
-        occ_coeffs = [intopt.sort_orbitals(occ_coeff, axis=[0]) for occ_coeff in occ_coeffs]
-        mo1s = [intopt.sort_orbitals(mo1, axis=[1]) for mo1 in mo1s]
-        cupy.cuda.get_current_stream().synchronize()
+        blksize = dfobj.get_blksize(mem_fraction=0.3)
+        if with_k:
+            vk = cupy.zeros_like(dms_3d)
+            mem_avail = get_avail_mem(exclude_memory_pool=True)
+            dm_batch_size = int(mem_avail * 0.8 / (blksize*nao*nocc * 8))
+            if factor_r is not None and factor_l.ndim == factor_r.ndim:
+                dm_batch_size = dm_batch_size // 2
+            dm_batch_size = min(dm_batch_size, n_dm)
+            assert dm_batch_size > 0
+            log.debug1('blksize=%d, dm_batch_size=%d', blksize, dm_batch_size)
 
-        futures = []
-        with ThreadPoolExecutor(max_workers=num_devices) as executor:
-            for device_id in range(num_devices):
-                future = executor.submit(
-                    _jk_task_with_mo1,
-                    dfobj, dms, mo1s, occ_coeffs,
-                    hermi=hermi, device_id=device_id,
-                    with_j=with_j, with_k=with_k)
-                futures.append(future)
+            if factor_r is None:
+                factor_l = factor_l.reshape(n_dm, nao, nocc)
+                buf = cp.empty(dm_batch_size * blksize * nao*nocc)
+            elif factor_l.ndim == factor_r.ndim:
+                factor_l = factor_l.reshape(n_dm, nao, nocc)
+                factor_r = factor_r.reshape(n_dm, nao, nocc)
+                buf = cp.empty(2 * dm_batch_size * blksize * nao*nocc)
+            elif factor_l.ndim < factor_r.ndim:
+                factor_l = factor_l.reshape(1, nao, nocc)
+                factor_r = factor_r.reshape(n_dm, nao, nocc)
+                buf = cp.empty(dm_batch_size+1, blksize * nao*nocc)
+                buf1 = buf[-1]
+            else:
+                factor_l = factor_l.reshape(n_dm, nao, nocc)
+                factor_r = factor_r.reshape(1, nao, nocc)
+                buf = cp.empty(dm_batch_size+1, blksize * nao*nocc)
+                buf1 = buf[-1]
 
-    elif hasattr(dms_tag, 'factor_l'):
-        futures = []
-        with ThreadPoolExecutor(max_workers=num_devices) as executor:
-            for device_id in range(num_devices):
-                future = executor.submit(
-                    _jk_via_decomposed_dm, dfobj, dms_tag,
-                    hermi=hermi, with_j=with_j, with_k=with_k,
-                    device_id=device_id)
-                futures.append(future)
+        for cderi, cderi_sparse in dfobj.loop(blksize=blksize, unpack=with_k):
+            if with_j:
+                vj_sparse += dm_sparse.dot(cderi_sparse.T).dot(cderi_sparse)
 
-    # general K matrix with density matrix
-    else:
-        dms = intopt.sort_orbitals(dms, axis=[1,2])
-        cupy.cuda.Stream.null.synchronize()
-        futures = []
-        with ThreadPoolExecutor(max_workers=num_devices) as executor:
-            for device_id in range(num_devices):
-                future = executor.submit(
-                    _jk_task_with_dm, dfobj, dms,
-                    hermi=hermi, device_id=device_id,
-                    with_j=with_j, with_k=with_k)
-                futures.append(future)
+            if with_k:
+                nL = len(cderi)
+                if factor_r is None:
+                    for i0, i1 in lib.prange(0, n_dm, dm_batch_size):
+                        rhok = ndarray((i1-i0,nao,nocc,nL), buffer=buf)
+                        contract('Lij,njk->nikL', cderi, factor_l[i0:i1], out=rhok)
+                        contract('nikL,njkL->nij', rhok, rhok, beta=1, out=vk[i0:i1])
+                elif factor_l.ndim == factor_r.ndim:
+                    for i0, i1 in lib.prange(0, n_dm, dm_batch_size):
+                        rhok, rhok1 = ndarray((2,i1-i0,nao,nocc,nL), buffer=buf)
+                        contract('Lij,njk->nikL', cderi, factor_l[i0:i1], out=rhok)
+                        contract('Lij,njk->nikL', cderi, factor_r[i0:i1], out=rhok1)
+                        contract('nikL,njkL->nij', rhok, rhok1, beta=1, out=vk[i0:i1])
+                elif factor_l.ndim < factor_r.ndim:
+                    rhok = ndarray((nao,nocc,nL), buffer=buf1)
+                    contract('Lij,jk->ikL', cderi, factor_l, out=rhok)
+                    for i0, i1 in lib.prange(0, n_dm, dm_batch_size):
+                        rhok1 = ndarray((i1-i0,nao,nocc,nL), buffer=buf)
+                        contract('Lij,njk->nikL', cderi, factor_r[i0:i1], out=rhok1)
+                        contract('nikL,jkL->nij', rhok, rhok1, beta=1, out=vk[i0:i1])
+                else:
+                    rhok1 = ndarray((nao,nocc,nL), buffer=buf1)
+                    contract('Lij,jk->ikL', cderi, factor_r, out=rhok1)
+                    for i0, i1 in lib.prange(0, n_dm, dm_batch_size):
+                        rhok = ndarray((i1-i0,nao,nocc,nL), buffer=buf)
+                        contract('Lij,njk->nikL', cderi, factor_l[i0:i1], out=rhok)
+                        contract('nikL,jkL->nij', rhok, rhok1, beta=1, out=vk[i0:i1])
+                rhok1 = rhok = None
+        if with_j:
+            vj = cp.zeros(dms_3d.shape)
+            vj[:,cols,rows] = vj_sparse
+            vj[:,rows,cols] = vj_sparse
+        return vj, vk
+
+    results = multi_gpu.run(proc, non_blocking=True)
 
     vj = vk = None
     if with_j:
-        vj = [future.result()[0] for future in futures]
-        vj = reduce_to_device(vj, inplace=True)
-        vj = intopt.unsort_orbitals(vj, axis=[1,2])
-        vj = vj.reshape(out_shape)
+        vj = multi_gpu.array_reduce([x[0] for x in results], inplace=True)
+        vj = vj.reshape(dms.shape)
+        if not out_cupy: vj = vj.get()
 
     if with_k:
-        vk = [future.result()[1] for future in futures]
-        vk = reduce_to_device(vk, inplace=True)
-        vk = intopt.unsort_orbitals(vk, axis=[1,2])
-        vk = vk.reshape(out_shape)
-
+        vk = multi_gpu.array_reduce([x[1] for x in results], inplace=True)
+        if symmetrize != 0:
+            vk = transpose_sum(vk, hermi=symmetrize)
+        vk = vk.reshape(dms.shape)
+        if not out_cupy: vk = vk.get()
     t1 = log.timer_debug1('vj and vk', *t1)
-    if out_cupy:
-        return vj, vk
-    else:
-        if vj is not None:
-            vj = vj.get()
-        if vk is not None:
-            vk = vk.get()
-        return vj, vk
+    return vj, vk
 
-def get_j(dfobj, dm, hermi=1, direct_scf_tol=1e-13):
-    intopt = getattr(dfobj, 'intopt', None)
-    if intopt is None:
-        dfobj.build(direct_scf_tol=direct_scf_tol)
-        intopt = dfobj.intopt
-    j2c = dfobj.j2c
-    rhoj = int3c2e.get_j_int3c2e_pass1(intopt, dm)
-    if dfobj.cd_low.tag == 'eig':
-        rhoj, _, _, _ = cupy.linalg.lstsq(j2c, rhoj)
-    else:
-        rhoj = cupy.linalg.solve(j2c, rhoj)
+def get_j(dfobj, dm, hermi=1):
+    from cupyx.scipy.linalg import solve_triangular
+    from gpu4pyscf.df.int3c2e_bdiv import int2c2e
 
-    rhoj *= 2.0
-    vj = int3c2e.get_j_int3c2e_pass2(intopt, rhoj)
+    if dfobj._cderi is not None:
+        return dfobj.get_jk(dm, hermi, with_k=False)
+
+    if dfobj.intopt is None:
+        dfobj.build(build_cderi=False)
+        j2c = int2c2e(dfobj.auxmol)
+        try:
+            dfobj.cd_j2c = cholesky(j2c)
+            dfobj.cd_tag = 'cd'
+        except RuntimeError:
+            dfobj.cd_j2c = j2c
+            dfobj.cd_tag = 'ed'
+
+    intopt = dfobj.intopt
+    mol = intopt.mol
+    auxmol = intopt.auxmol
+    dm = mol.apply_C_mat_CT(dm)
+    rhoj = intopt.contract_dm(dm, hermi=hermi)
+    if dfobj.cd_tag == 'ed':
+        rhoj = cp.linalg.solve(dfobj.cd_j2c, rhoj.T)[0]
+    else:
+        rhoj = solve_triangular(dfobj.cd_j2c, rhoj.T, lower=True)
+        rhoj = solve_triangular(dfobj.cd_j2c.T, rhoj, lower=False)
+
+    auxvec = auxmol.C_dot_mat(rhoj.T)
+    vj = intopt.contract_auxvec(auxvec)
     return vj
 
 density_fit = _density_fit
@@ -887,25 +638,10 @@ def decompose_rdm1_svd(dm, hermi=0):
         return u[:,:,mask], contract('si,sip->spi', s[:,mask], vh[:,mask])
 
 def _make_factorized_dm(factor_l, factor_r, symmetrize=1):
-    if factor_r.ndim == 2:
-        dm = factor_l.dot(factor_r.T)
-    elif factor_l.ndim == 2:
-        dm = contract('pi,xqi->xpq', factor_l, factor_r)
-    elif factor_l.ndim == 3:
-        dm = contract('xpi,xqi->xpq', factor_l, factor_r)
-    else:
-        raise RuntimeError(f'{factor_l.shape} not supported')
-
-    if dm.ndim == 2:
-        if symmetrize == 1:
-            dm = dm + dm.T
-        elif symmetrize == 2:
-            dm = dm - dm.T
-    else:
-        if symmetrize == 1:
-            dm = dm + dm.transpose(0,2,1)
-        elif symmetrize == 2:
-            dm = dm - dm.transpose(0,2,1)
+    dm = cp.matmul(factor_l, factor_r.T)
+    if symmetrize == 1 or symmetrize == 2:
+        nao = dm.shape[-1] # dm1 may have dimensions > 3
+        transpose_sum(dm.reshape(-1,nao,nao), inplace=True, hermi=symmetrize)
     return tag_array(dm, factor_l=factor_l, factor_r=factor_r, symmetrize=symmetrize)
 
 def _tag_factorize_dm(dm, hermi=0):
