+++ gpu4pyscf/df/hessian/uks.py
@@ -26,8 +26,8 @@ from gpu4pyscf.grad import rhf as rhf_grad
 from gpu4pyscf.hessian import rhf as rhf_hess
 from gpu4pyscf.hessian import uhf as uhf_hess
 from gpu4pyscf.hessian import uks as uks_hess
+from gpu4pyscf.df.hessian import rhf as df_rhf_hess
 from gpu4pyscf.df.hessian import uhf as df_uhf_hess
-from gpu4pyscf.df.hessian.uhf import _partial_hess_ejk, _get_jk_ip
 from gpu4pyscf.hessian.rks import _get_enlc_deriv2, _get_vnlc_deriv1
 from gpu4pyscf.lib import logger
 from gpu4pyscf.lib.cupy_helper import contract
@@ -42,30 +42,29 @@ def partial_hess_elec(hessobj, mo_energy=None, mo_coeff=None, mo_occ=None,
     if mo_energy is None: mo_energy = mf.mo_energy
     if mo_occ is None:    mo_occ = mf.mo_occ
     if mo_coeff is None:  mo_coeff = mf.mo_coeff
-    if atmlst is None: atmlst = range(mol.natm)
 
-    mocca = mo_coeff[0][:,mo_occ[0]>0]
-    moccb = mo_coeff[1][:,mo_occ[1]>0]
-    dm0a = mocca.dot(mocca.T)
-    dm0b = moccb.dot(moccb.T)
+    if mf.with_df.intopt is None:
+        mf.with_df.build(build_cderi=False)
+    intopt = mf.with_df.intopt
+    mf.with_df.reset() # Release GPU memory
 
     omega, alpha, hyb = mf._numint.rsh_and_hybrid_coeff(mf.xc, spin=mol.spin)
-    with_k = mf._numint.libxc.is_hybrid_xc(mf.xc)
-    de2, ej, ek = _partial_hess_ejk(hessobj, mo_energy, mo_coeff, mo_occ,
-                                    atmlst, max_memory, verbose,
-                                    with_j=True, with_k=with_k)
-    de2 += ej  # (A,B,dR_A,dR_B)
-    if with_k:
-        de2 -= hyb * ek
+    dm0 = mf.make_rdm1(mo_coeff, mo_occ)
+    de2 = df_uhf_hess._jk_energy_per_atom(intopt, dm0, 1., hyb, verbose=log)
 
     if abs(omega) > 1e-10 and abs(alpha-hyb) > 1e-10:
-        ek_lr = _partial_hess_ejk(hessobj, mo_energy, mo_coeff, mo_occ,
-                                  atmlst, max_memory, verbose,
-                                  with_j=False, with_k=True, omega=omega)[2]
-        de2 -= (alpha - hyb) * ek_lr
+        de2 += df_uhf_hess._jk_energy_per_atom(intopt, dm0, 0., alpha-hyb,
+                                               omega, verbose=log)
+    t1 = log.timer_debug1('computing ej, ek', *t1)
+
+    # Energy weighted density matrix
+    dme0_sf  = (mo_coeff[0] * (mo_occ[0]*mo_energy[0])).dot(mo_coeff[0].T)
+    dme0_sf += (mo_coeff[1] * (mo_occ[1]*mo_energy[1])).dot(mo_coeff[1].T)
+    dm0_sf = dm0[0] + dm0[1]
+    de2 += df_uhf_hess._hcore_energy(hessobj, dm0_sf, dme0_sf)
+    log.timer_debug1('hcore contribution', *t1)
 
     max_memory = None
-    t1 = log.timer_debug1('computing ej, ek', *t1)
     de2 += uks_hess._get_exc_deriv2(hessobj, mo_coeff, mo_occ, (dm0a, dm0b), max_memory)
     if mf.do_nlc():
         de2 += _get_enlc_deriv2(hessobj, mo_coeff, mo_occ, max_memory)
@@ -78,32 +77,22 @@ def make_h1(hessobj, mo_coeff, mo_occ, chkfile=None, atmlst=None, verbose=None):
     natm = mol.natm
     assert atmlst is None or atmlst ==range(natm)
     mf = hessobj.base
+    if mf.with_df.intopt is None:
+        mf.with_df.build(build_cderi=False)
+    intopt = mf.with_df.intopt
+
     ni = mf._numint
     ni.libxc.test_deriv_order(mf.xc, 2, raise_error=True)
     omega, alpha, hyb = ni.rsh_and_hybrid_coeff(mf.xc, spin=mol.spin)
     mem_now = lib.current_memory()[0]
     max_memory = max(2000, mf.max_memory*.9-mem_now)
 
-    with_k = ni.libxc.is_hybrid_xc(mf.xc)
-
-    vj1, vk1 = _get_jk_ip(hessobj, mo_coeff, mo_occ, chkfile,
-                          atmlst, verbose, with_j=True, with_k=True)
-    vj1a, vj1b = vj1
-    h1moa = vj1a
-    h1mob = vj1b
-
-    if with_k:
-        vk1a, vk1b = vk1
-        h1moa -= hyb * vk1a
-        h1mob -= hyb * vk1b
-    vj1 = vk1 = vj1a = vj1b = vk1a = vk1b = None
+    h1mo = df_uhf_hess._get_veff(intopt, mo_coeff, mo_occ, 1., hyb)
 
     if abs(omega) > 1e-10 and abs(alpha-hyb) > 1e-10:
-        _, vk1_lr = _get_jk_ip(hessobj, mo_coeff, mo_occ, chkfile,
-                               atmlst, verbose, with_j=False, with_k=True, omega=omega)
-        vk1a, vk1b = vk1_lr
-        h1moa -= (alpha - hyb) * vk1a
-        h1mob -= (alpha - hyb) * vk1b
+        veff_lr = df_uhf_hess._get_veff(intopt, mo_coeff, mo_occ, 0., alpha-hyb, omega)
+        h1mo[0] += veff_lr[0]
+        h1mo[1] += veff_lr[1]
 
     gobj = hessobj.base.nuc_grad_method()
     h1moa += rhf_grad.get_grad_hcore(gobj, mo_coeff[0], mo_occ[0])
@@ -128,5 +117,5 @@ class Hessian(uks_hess.Hessian):
     auxbasis_response = 2
     partial_hess_elec = partial_hess_elec
     make_h1 = make_h1
-    get_jk_mo = df_uhf_hess._get_jk_mo
+    get_jk_mo = df_rhf_hess._get_jk_mo
     to_cpu = df_uhf_hess.Hessian.to_cpu
